rest/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-filter==2.2.0
djangorestframework==3.10.1
idna==2.8
Markdown==3.1.1
Pillow==6.1.0
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
wincertstore==0.2
